PR changes based on code review
