+++ strr-api/src/strr_api/validators/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module is for business logic validation.
+"""
